pyproject.toml
[project]
name = "datascope"
version = "0.1.0"
description = "Data Debugging Agent for Databricks"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Databricks
    "databricks-sdk>=0.20.0",
    "databricks-mcp>=0.1.0",
    "databricks-agents>=1.0.0",
    "mcp>=1.9.0",
    # LangGraph + LangChain
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=0.1.0",  # Lakebase state management
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",  # For External Endpoint (OpenAI-compatible API)
    # Core
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    # MLflow
    "mlflow>=3.1.0",
    # UI
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]
ui = [
    "streamlit>=1.35.0",
    "streamlit-chat>=0.1.1",
]

[project.scripts]
datascope = "datascope.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/datascope"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
